+++ apps/watch/src/components/Video/index.ts
@@ -1,0 +1,1 @@
+export { VideoCard } from './Card/VideoCard'
