--- apps/watch/src/components/Videos/VideoInfo/VideoInfo.tsx
@@ -1,5 +1,0 @@
-import { ReactElement } from 'react'
-
-export function VideoInfo(): ReactElement {
-  return <div>VideoInfo</div>
-}
