+++ apps/watch/src/components/Videos/index.ts
@@ -1,0 +1,1 @@
+export { Videos } from './Videos'
